fix: shell=True seems to be broken on non-Windows
With shell=True, it's better to provide a str rather than list[str].

From docs: "On POSIX with shell=True, (...) If args is a sequence, the first item specifies the command string, and any additional items will be treated as additional arguments to the shell itself.". On Windows there's automatic conversion, which might be undesired.

@see https://stackoverflow.com/questions/3172470/actual-meaning-of-shell-true-in-subprocess

Signed-off-by: Jack Cherng <jfcherng@gmail.com>