feat: add "blender" and "gdb" dev_environments
These two environments will query the additional paths by calling
the Blender or gdb executable. More testing is probably needed for
Blender on macOS, because it's probably not on the $PATH, and the
actual CLI executable is inside the /Applications/Blender.app folder.

What's also not great right now is that both these environments have a compiled
native extension (_bpy and _gdb, respectively) and both of these applications
don't have a corresponding .pyi interface file for these native extensions.

Actually, the same problem occurs for Sublime Text (sublime_api module is a
native extension). However in the case of Sublime Text, this sublime_api module
is completely wrapped in pure python classes/functions. So we don't notice that
it doesn't have a .pyi interface file.